requirements.txt
Django>=4.2,<5
psycopg2-binary>=2.9
gunicorn>=20.1
python-dotenv>=1.0
django-storages[boto3]>=1.13
boto3>=1.26
djangorestframework>=3.14
django-cors-headers>=3.15
watchdog>=2.1
google-cloud-firestore>=2.11
